音乐随动

+++ app/src/main/java/com/morrigan/m/ble/BleController.java
@@ -202,7 +202,7 @@ public class BleController extends AbstractBleController {
         massageTask.executeOnExecutor(EXECUTOR_SERVICE_SINGLE);
     }
 
-    public void musicMassageAsync(byte decibel) {
+    public void musicMassageAsync(int decibel) {
         MassageTask massageTask = new MassageTask(new MassageData(true, decibel));
         massageTask.executeOnExecutor(EXECUTOR_SERVICE_SINGLE);
     }
